Ear Protection File Manager / RPC Done|Working
Reverted | Changes to the Constants(3_Game) pathing like that wasn't needed.
Reverted | Changes to the File Manager(3_Game), This file was working just needed so extra steps down the road.
New        | Added a RPCManager so that the server can tell client the information on the file located on the server.
Changed | Removed the Static array of items from PlayerBase.c and implemented a TStringarray for the RPCManager to store items.